Add EquivalenceSummary and (de)serialization configuration
EquivalenceSummary records computed candidates and resultant
equivalents of the equivalence process. This intended for use in other
equivalence settings, i.e. the results of a Brand's children can
influence the result of the Brand itself, or vice-versa.

Any other use, such as resolving content, is probably not a good idea.