BT TV Channels publisher enabled without an API key
For backward-compatibility, this must be enabled without an API
key, since some queries are being done without an API key, which
have worked until now because the channel_group_summary annotation
did not filter according to enabled publishers.

It now does, so until these queries have an API key added, we are
making this source enabled without an API key.