Added customFields map field to Identified model (#355)
* Added customFields map field to Identified model

The field is intended to be used for arbitrary data
that does not fit within the confines of the regular
atlas model e.g. additional title fields or customer-specific
fields.

Added merging logic for customFields which will
merge fields from each map into the highest-precedence
customFields map where the field does not already exist.
This is to mimic behaviour of merging logic of regular
fields.

Added a custom_fields annotation which will be used to output
the field.