allows for ordering of elements of set fields
We're experiencing an issue where writing the same item
through owl-api via owl-client results in changes being
detected because the elements of a set field e.g. certificates
are in a different order. This results in writing pointlessly
to db and also adding to the queue, in some cases causing
huge queues (e.g. ingest-imdb)